Rename SearchField to SearchForm
