+++ src/app/search/page.tsx
@@ -2,7 +2,7 @@ import clsx from 'clsx';
 import { type Metadata } from 'next';
 import { type ReactNode } from 'react';
 
-import { PostList, SearchField } from '@/components';
+import { PostList, SearchForm } from '@/components';
 
 export const metadata: Metadata = {
   title: 'Search',
@@ -17,7 +17,7 @@ const Page = (): ReactNode => {
 
   return (
     <main className={classNames}>
-      <SearchField />
+      <SearchForm />
       <PostList />
     </main>
   );
