--- src/components/SearchField/index.ts
@@ -1,1 +1,0 @@
-export { default as SearchField } from './SearchField';
